Init with functional plugin
This version uses the local plugin feature of Gatsby by placing the
would-be npm package inside the plugins folder. I'm working on a proper
npm version of the plugin, but this commit serves to test that the code
for the plugin works. Run gatsby develop and the LaTeX to SVG conversion
should work on src/pages/index.md